requirements.txt
fastapi
uvicorn
pytesseract
Pillow
langchain
openai
langchain-community
python-multipart 